Split treasury source context tests
